VRG: Merge status PVC routines to one file
