+++ controllers/vrg_status_pvcs.go
@@ -49,3 +49,33 @@ func FindProtectedPvcAndIndex(
 
 	return nil, len(vrg.Status.ProtectedPVCs)
 }
+
+func (v *VRGInstance) findFirstProtectedPVCWithName(pvcName string) *ramen.ProtectedPVC {
+	for index := range v.instance.Status.ProtectedPVCs {
+		protectedPVC := &v.instance.Status.ProtectedPVCs[index]
+		if protectedPVC.Name == pvcName {
+			return protectedPVC
+		}
+	}
+
+	return nil
+}
+
+func (v *VRGInstance) vrgStatusPvcNamespacesSetIfUnset() {
+	vrg := v.instance
+
+	for i := range vrg.Status.ProtectedPVCs {
+		pvc := &vrg.Status.ProtectedPVCs[i]
+		log := v.log.WithValues("PVC", pvc.Name)
+
+		if pvc.Namespace != "" {
+			log.V(1).Info("VRG status PVC namespace set already", "namespace", pvc.Namespace)
+
+			continue
+		}
+
+		log.Info("VRG status PVC namespace unset; setting")
+
+		pvc.Namespace = vrg.GetNamespace()
+	}
+}
